requirements.txt
pyjwt
paho-mqtt
cryptography